fix(egress): explicit int casts in reserve refresh due-check SQL
Cast the interval-minutes expression and the $1 freshness param to int so
LEAST()/interval building can't hit a type-inference error that would
silently break the subscription refresher (and thus the reserve keep-warm).

+++ apps/backend/src/operations/operations.service.ts
@@ -1621,7 +1621,7 @@ export class OperationsService {
         WHERE enabled = true
           AND (last_fetched_at IS NULL
                OR last_fetched_at < now()
-                  - (LEAST(COALESCE(update_interval_hours, 12) * 60, $1)::text || ' minutes')::interval)
+                  - (LEAST((COALESCE(update_interval_hours, 12) * 60)::int, $1::int)::text || ' minutes')::interval)
       `,
       [cap],
     );
